+++ packages/backend/src/secret.ts
@@ -1,6 +1,5 @@
 import { BackendSecret } from '@aws-amplify/plugin-types';
 import { CfnTokenBackendSecret } from './engine/backend-secret/backend_secret.js';
-import { getSecretClient } from '@aws-amplify/backend-secret';
 import { BackendSecretFetcherProviderFactory } from './engine/backend-secret/backend_secret_fetcher_provider_factory.js';
 import { BackendSecretFetcherFactory } from './engine/backend-secret/backend_secret_fetcher_factory.js';
 
@@ -8,9 +7,7 @@ import { BackendSecretFetcherFactory } from './engine/backend-secret/backend_sec
  * Factory function for initializing a backend secret.
  */
 export const secret = (name: string): BackendSecret => {
-  const secretProviderFactory = new BackendSecretFetcherProviderFactory(
-    getSecretClient()
-  );
+  const secretProviderFactory = new BackendSecretFetcherProviderFactory();
   const secretResourceFactory = new BackendSecretFetcherFactory(
     secretProviderFactory
   );